+++ src/flexpath.cpp
@@ -937,11 +937,11 @@ ErrorCode FlexPath::to_oas(OasisStream& out, OasisState& state) {
     return error_code;
 }
 
-ErrorCode FlexPath::to_svg(FILE* out, double scaling) {
+ErrorCode FlexPath::to_svg(FILE* out, double scaling, uint32_t precision) {
     Array<Polygon*> array = {0};
     ErrorCode error_code = to_polygons(array);
     for (uint64_t i = 0; i < array.count; i++) {
-        ErrorCode err = array[i]->to_svg(out, scaling);
+        ErrorCode err = array[i]->to_svg(out, scaling, precision);
         if (err != ErrorCode::NoError) error_code = err;
         array[i]->clear();
         free_allocation(array[i]);