+++ src/label.h
@@ -59,7 +59,7 @@ struct Label {
     // These functions output the label in the GDSII and SVG formats.  They are
     // not supposed to be called by the user.
     ErrorCode to_gds(FILE* out, double scaling) const;
-    ErrorCode to_svg(FILE* out, double scaling) const;
+    ErrorCode to_svg(FILE* out, double scaling, uint32_t precision) const;
 };
 
 }  // namespace gdstk