+++ src/robustpath.h
@@ -257,7 +257,7 @@ struct RobustPath {
     // valid for to_oas, even though no fracturing ever occurs for OASIS files.
     ErrorCode to_gds(FILE* out, double scaling) const;
     ErrorCode to_oas(OasisStream& out, OasisState& state) const;
-    ErrorCode to_svg(FILE* out, double scaling) const;
+    ErrorCode to_svg(FILE* out, double scaling, uint32_t precision) const;
 
    private:
     void simple_scale(double scale);